resolve comments
* move methods out from authService
* expose updateAuthState from oktaAuth
* push task into promiseQ inside updateAuthState
